requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
gunicorn
django-model-utils
django-ses
django-import-export
requests==2.22.0
boto3==1.9.156
tinys3==0.1.12
Pillow==6.0.0
python-resize-image==1.1.18
